Introduced generic parent class
Using a generic parent class to implement the non-specific image
processing is a more elegant way of doing things, and it makes it easier
to create future sub-classes for different kinds of input (e.g. a series
of images, or a video, or a different type of camera hardware/software).