Clean joining of Threads and Processes.
This prevents leaving them running in the background. The change is
implemented in a slightly more elegant way of allowing a user-defined
way of closing the connection, without users having to explicitly call
the generic.EyeTracker's closing implementation. EyeTracker.close now
calls EyeTracker._close, which is intended to be overwritten as needed
in a user's sub-class (see webcam.WebCamTracker for an example).