Added an ImageTracker class, and several updates
The largest change is the addition of the ImageTracker class, which
allows for eye tracking in a directory of existing images.

Other changes include the eye detection now being optional, and the
fixing of DEBUG things. (You can now run the modules to test them out,
and the DEBUG images will also mark the glint.)